Added useSearchBook hook and comments
Implemented useSearchBook hook for fetching the API endpoint to search book by title and author and return if it exists or not. The hook is used in the useAddBook hook, so now a book only added if it does not exits in the database. Additionally, I commented both the useAddBook and useSearchBook hooks.