Refactored the code to use local states instead of context
The code was refactored to utilise local states instead of relying on the book and listing context and reducer. This decision was made because, in this specific scenario, using context introduced unnecessary complexity to the code. I also renamed the utils folder to hooks, because only hooks were added there.